Divided CameraScreen code in two parts as UI and Backend
